seperate presets out from application state
