+++ frontend/src/Components/BarChart/ExtraPairOutcomes.tsx
@@ -3,9 +3,9 @@ import Store from "../../Interfaces/Store";
 import styled from "styled-components";
 import { inject, observer } from "mobx-react";
 import { ScaleBand, scaleOrdinal, range, scaleLinear, ScaleOrdinal, max, format, interpolateGreys, mean, scaleBand } from "d3";
-import { extraPairWidth } from "../../Interfaces/ApplicationState"
+import { extraPairWidth } from "../../PresetsProfile"
 import { Popup } from "semantic-ui-react";
-import { secondary_gray, greyScaleRange } from "../../ColorProfile";
+import { secondary_gray, greyScaleRange } from "../../PresetsProfile";
 
 interface OwnProps {
     outcomeName: string;
