+++ frontend/src/Components/InterventionPlot/ExtraPairBaiscInt.tsx
@@ -3,9 +3,9 @@ import Store from "../../Interfaces/Store";
 import styled from "styled-components";
 import { inject, observer } from "mobx-react";
 import { ScaleBand, scaleOrdinal, range, scaleLinear, ScaleOrdinal, max, format, interpolateGreys, scaleBand } from "d3";
-import { extraPairWidth } from "../../Interfaces/ApplicationState"
+import { extraPairWidth } from "../../PresetsProfile"
 import { Popup } from "semantic-ui-react";
-import { secondary_gray } from "../../ColorProfile";
+import { secondary_gray } from "../../PresetsProfile";
 
 interface OwnProps {
     totalData: any[];
