+++ frontend/src/Components/InterventionPlot/ExtraPairViolinInt.tsx
@@ -3,8 +3,8 @@ import Store from "../../Interfaces/Store";
 import styled from "styled-components";
 import { inject, observer } from "mobx-react";
 import { ScaleBand, scaleOrdinal, range, scaleLinear, ScaleOrdinal, line, curveCatmullRom, median, format, scaleBand } from "d3";
-import { extraPairWidth, offset } from "../../Interfaces/ApplicationState"
-import { preop_color, postop_color, basic_gray } from "../../ColorProfile";
+import { extraPairWidth } from "../../PresetsProfile"
+import { preop_color, postop_color, basic_gray } from "../../PresetsProfile";
 import { create as createpd } from "pdfast";
 import { Popup } from "semantic-ui-react";
