+++ frontend/src/Components/PatientComparisonWrapper.tsx
@@ -7,7 +7,7 @@ import styled from "styled-components";
 import { inject, observer } from "mobx-react";
 import Store from "../Interfaces/Store";
 import { Message, List } from "semantic-ui-react";
-import { HIPAA_Sensitive, AxisLabelDict, stateUpdateWrapperUseJSON } from "../Interfaces/ApplicationState";
+import { HIPAA_Sensitive, AxisLabelDict, stateUpdateWrapperUseJSON } from "../PresetsProfile";
 
 interface OwnProps {
     store?: Store;
